requirements.txt
easyocr==1.7.1
numpy==1.26.1
Pillow==10.1.0
streamlit==1.28.1
